Fixed bug in merging
The merging when reqPercent is greater than 0 had a bug where a floating point number was expected, but two integers were used in a division. The result of the division was an integer (0 or 1) instead of a floating point number.

+++ src/Cluster.hpp
@@ -185,7 +185,7 @@ public:
         }
       }
 
-      if (possibleMergePoints / toMergeCluster.points.size() >= reqPercent) {
+      if ((double) possibleMergePoints / toMergeCluster.points.size() >= reqPercent) {
         for (int i = 0; i < toMergeCluster.points.size(); i++) {
 
           Add(toMergeCluster.pointsReal[i],
